requirements.txt
aiogram==3.21.0
sqlalchemy==1.4.49
python-dotenv==1.0.1
aiosqlite==0.19.0
alembic==1.12.1
g4f==0.1.9.0
asyncio==3.4.3
pydantic==2.5.2
typing-extensions==4.8.0
pytz==2023.3
requests==2.31.0
loguru==0.7.2